discord-mcp-server/discord_mcp_server.py
from datetime import timedelta
from fastmcp import FastMCP
import discord

# Initialize Discord client
discord_client = discord.Client(intents=discord.Intents.all() )

# Initialize FastMCP server
mcp = FastMCP("Discord MCP Server")

@mcp.tool
async def send_message(channel_id: int, message: str) -> str:
    """
    Send a message to a Discord channel.
    
    :param channel_id: The ID of the Discord channel to send the message to
    :param message: The message content to send
    :return: Success message
    """
    try:
        return "Message sent to channel"
        # channel = discord_client.get_channel(channel_id)
        # if channel:
        #     await channel.send(message)
        #     return f"Message sent to channel {channel_id}"
        # else:
        #     return f"Could not find channel with ID {channel_id}"
    except Exception as e:
        return f"Error sending message: {str(e)}"
        
@mcp.tool
async def get_messages(channel_id: int, limit: int = 10) -> list:
    """
    Retrieve message history from a Discord channel.
    
    :param channel_id: The ID of the Discord channel to retrieve messages from
    :param limit: Number of messages to retrieve (default 10)
    :return: List of message contents
    """
    try:
        return ["Messages retrieved"]
        # channel = discord_client.get_channel(channel_id)
        # if channel:
        #     messages = [message.content async for message in channel.history(limit=limit)]
        #     return messages
        # else:
        #     return [f"Could not find channel with ID {channel_id}"]
    except Exception as e:
        return [f"Error retrieving messages: {str(e)}"]
        
@mcp.tool
async def get_channel_info(channel_id: int) -> dict:
    """
    Fetch metadata about a Discord channel.
    
    :param channel_id: The ID of the Discord channel to fetch information about
    :return: Channel metadata
    """
    try:
        return {"Channel info fetched": "Channel info fetched"}
        # channel = discord_client.get_channel(channel_id)
        # if channel:
        #     return {
        #         "name": channel.name,
        #         "type": channel.type,
        #         "topic": channel.topic,
        #         "is_nsfw": channel.is_nsfw(),
        #         "is_private": channel.is_private(),
        #         "is_nsfw": channel.is_nsfw(),
        #         "is_private": channel.is_private(),
        #         "is_nsfw": channel.is_nsfw(),
        #         "is_private": channel.is_private(),
        #     }
        # else:
        #     return {"error": f"Could not find channel with ID {channel_id}"}
    except Exception as e:
        return {"error": f"Error fetching channel info: {str(e)}"}
        
@mcp.tool
async def search_messages(channel_id: int, query: str, limit: int = 10) -> list:
    """
    Search for messages in a Discord channel.
    
    :param channel_id: The ID of the Discord channel to search in
    :param query: The search query
    :param limit: Number of messages to retrieve (default 10)
    :return: List of matching message contents
    """
    try:
        return ["Messages searched"]
        # channel = discord_client.get_channel(channel_id)
        # if channel:
        #     messages = [message.content async for message in channel.history(limit=limit)]
        #     return [message for message in messages if query in message]
        # else:
        #     return [f"Could not find channel with ID {channel_id}"]
    except Exception as e:
        return [f"Error searching messages: {str(e)}"]
        
@mcp.tool
async def moderate_content(channel_id: int, message_id: int, action: str) -> str:
    """
    Moderate content in a Discord channel.
    
    :param channel_id: The ID of the Discord channel to moderate
    :param message_id: The ID of the message to moderate
    :param action: The action to take (e.g., "delete", "warn", "mute")
    :return: Success message
    """
    try:
        return "Content moderated"
        # channel = discord_client.get_channel(channel_id)
        # if channel:
        #     message = await channel.fetch_message(message_id)
        #     if action == "delete":
        #         await message.delete()
        #         return f"Message {message_id} deleted"
        #     elif action == "warn":
        #         await message.reply("This message has been flagged as inappropriate.")
        #         return f"Message {message_id} flagged as inappropriate"
        #     elif action == "mute":
        #         await message.author.timeout(duration=timedelta(minutes=10))
        #         return f"Message {message_id} muted for 10 minutes"
        #     else:
        #         return f"Invalid action: {action}"
    except Exception as e:
        return f"Error moderating content: {str(e)}"

# Run the MCP server
if __name__ == "__main__":
    mcp.run()
