discord-mcp-server/requirements.txt
fastapi
uvicorn
discord.py
pydantic
sqlalchemy
aiohttp
pytest
fastmcp
python-dotenv
streamlit
